packages/document-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
markdown==3.5.1
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
openai==1.3.8
chromadb==0.4.18
pinecone-client==2.2.4
weaviate-client==3.25.3
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
pyyaml==6.0.1
chardet==5.2.0
numpy==1.24.4
beautifulsoup4==4.12.2 